factor out ScraperHealthMonitor
