+++ kifi-backend/modules/shoebox/app/com/keepit/shoebox/ShoeboxModule.scala
@@ -6,7 +6,7 @@ import com.keepit.social.SecureSocialModule
 import com.keepit.common.mail.MailModule
 import com.keepit.common.analytics.AnalyticsModule
 import com.keepit.model.ProdSliderHistoryTrackerModule
-import com.keepit.scraper.{ ScraperServiceClientModule, ScrapeSchedulerModule, ProdScraperServiceClientModule }
+import com.keepit.scraper.{ ScraperServiceClientModule, ScraperHealthMonitorModule, ProdScraperServiceClientModule }
 import com.keepit.classify.DomainTagImporterModule
 import com.keepit.common.store.ShoeboxDevStoreModule
 import com.keepit.inject.{ CommonServiceModule, ConfigurationModule }
@@ -38,7 +38,7 @@ abstract class ShoeboxModule(
     val analyticsModule: AnalyticsModule,
     val domainTagImporterModule: DomainTagImporterModule,
     val cacheModule: ShoeboxCacheModule,
-    val scrapeSchedulerModule: ScrapeSchedulerModule,
+    val scraperHealthMonitorModule: ScraperHealthMonitorModule,
     val fjMonitorModule: ForkJoinContextMonitorModule,
     val externalServiceModule: ExternalServiceModule,
     val rekeepStatsUpdaterModule: ReKeepStatsUpdaterModule) extends ConfigurationModule with CommonServiceModule {
