+++ kifi-backend/modules/shoebox/test/com/keepit/commanders/CollectionCommanderTest.scala
@@ -6,7 +6,7 @@ import com.keepit.model._
 import com.keepit.heimdal.HeimdalContext
 import org.joda.time.DateTime
 import com.keepit.common.time._
-import com.keepit.scraper.FakeScrapeSchedulerModule
+import com.keepit.scraper.FakeScraperHealthMonitorModule
 import play.api.libs.json.Json
 import scala.Some
 import com.keepit.model.KeepToCollection
@@ -19,7 +19,7 @@ class CollectionCommanderTest extends Specification with ShoeboxTestInjector {
 
   implicit val context = HeimdalContext.empty
 
-  def modules = FakeScrapeSchedulerModule() :: FakeSearchServiceClientModule() :: Nil
+  def modules = FakeScraperHealthMonitorModule() :: FakeSearchServiceClientModule() :: Nil
 
   def prenormalize(url: String)(implicit injector: Injector): String = inject[NormalizationService].prenormalize(url).get
