+++ kifi-backend/modules/shoebox/test/com/keepit/controllers/admin/AdminDashboardControllerTest.scala
@@ -22,7 +22,7 @@ import com.keepit.common.net.FakeHttpClientModule
 import com.keepit.common.store.ShoeboxFakeStoreModule
 import com.keepit.common.healthcheck.FakeAirbrakeModule
 import com.keepit.search.TestSearchServiceClientModule
-import com.keepit.scraper.{ TestScraperServiceClientModule, FakeScrapeSchedulerModule }
+import com.keepit.scraper.{ TestScraperServiceClientModule, FakeScraperHealthMonitorModule }
 import com.keepit.common.external.FakeExternalServiceModule
 import com.keepit.cortex.FakeCortexServiceClientModule
 
@@ -30,7 +30,7 @@ class AdminDashboardControllerTest extends Specification with ShoeboxApplication
 
   def requiredModules = Seq(
     TestSearchServiceClientModule(),
-    FakeScrapeSchedulerModule(),
+    FakeScraperHealthMonitorModule(),
     ProdShoeboxSecureSocialModule(),
     FakeHttpClientModule(),
     ShoeboxFakeStoreModule(),
