+++ kifi-backend/modules/shoebox/test/com/keepit/shoebox/ShoeboxModuleTest.scala
@@ -25,7 +25,7 @@ import com.keepit.common.analytics.TestAnalyticsModule
 import com.keepit.model.TestSliderHistoryTrackerModule
 import com.keepit.classify.FakeDomainTagImporterModule
 import com.keepit.eliza.FakeElizaServiceClientModule
-import com.keepit.scraper.{ TestScrapeSchedulerConfigModule, TestScraperServiceClientModule, FakeScrapeSchedulerModule }
+import com.keepit.scraper.{ TestScrapeSchedulerConfigModule, TestScraperServiceClientModule, FakeScraperHealthMonitorModule }
 import com.keepit.common.healthcheck.FakeAirbrakeModule
 import com.keepit.heimdal.TestHeimdalServiceClientModule
 import com.keepit.abook.TestABookServiceClientModule
@@ -65,7 +65,7 @@ class ShoeboxModuleTest extends Specification with Logging with ShoeboxApplicati
         TestGraphServiceClientModule(),
         GeckoboardModule(),
         FakeShoeboxServiceModule(), // This one should not be required once the Scraper is off Shoebox
-        FakeScrapeSchedulerModule(), // This one should not be required once the Scraper is off Shoebox
+        FakeScraperHealthMonitorModule(),
         TestScrapeSchedulerConfigModule(),
         FakeElizaServiceClientModule(),
         FakeAirbrakeModule(),
